Commit 02/01/20
Tentative de création d'une fonction unique sur le controller backend pour appeler à la fois les posts et les commentaires en administrateur

+++ index.php
@@ -31,52 +31,37 @@ try {
                 }
                 break;
 
-           /** case('isAdmin'):
-                if (isset($_POST['password']) && $_POST['password'] ==  "pass"){
-                    $isAuthenticated = true;
-                }
-                elseif (!$_POST['password'] ==  "pass"){
-                    throw new Exception('Attention le mot de passe est incorrect !');
-                }
-                break; */
-
             case('admin'):
+                    if(!isset($_SESSION['admin'])) {
+                        isAdmin();
+                    } elseif(isset($_SESSION['admin'])){
+                            if(!isset($_GET['edit'])){
+                                admin(0);
+                            } elseif(isset($_POST['save'])){
+                                if (!empty($_POST['title']) && !empty($_POST['content'])) {
+                                    createPost($_POST['title'], $_POST['content']);
+                                    var_dump($_POST['title']);
+                                } else {
+                                    throw new Exception('Tous les champs ne sont pas remplis !');
+                                }
+                            } elseif (isset($_GET['edit'])){
+                                admin($_GET['edit']);
 
-                    if(!isset($_GET['edit'])){
-                        admin(0);
-
-                    } elseif(isset($_POST['save'])){
-                        if (!empty($_POST['title']) && !empty($_POST['content'])) {
-                            createPost($_POST['title'], $_POST['content']);
-                            var_dump($_POST['title']);
+                                if (isset($_POST['update'])) {
+                                    if (!empty($_POST['title']) && !empty($_POST['content'])) {
+                                        updatePost($_GET['edit'],$_POST['title'], $_POST['content']);
+                                    }
+                                    else {
+                                        throw new Exception('Tous les champs ne sont pas remplis !');
+                                    }
+                                }  elseif (isset($_POST['delete'])){
+                                    deletePost($_GET['edit']);
+                                    throw new Exception('Tous les champs ne sont pas remplis !');
+                                }
+                            }
                         } else {
-                            throw new Exception('Tous les champs ne sont pas remplis !');
-                        }
-                    } elseif (isset($_GET['edit'])){
-                        admin($_GET['edit']);
-
-                        if (isset($_POST['update'])) {
-                            if (!empty($_POST['title']) && !empty($_POST['content'])) {
-                                updatePost($_GET['edit'],$_POST['title'], $_POST['content']);
-                                /**$id = $_GET['edit'];
-                                $title = $_POST['title'];
-                                $content = $_POST['content'];
-                                $db = new \PDO('mysql:host=localhost:3308;dbname=oc4;charset=utf8', 'root', 'root');
-                                $update = $db->prepare("UPDATE posts SET title= :title, content=:content WHERE id=:id" );
-                                $update->execute(array(':id'=>$id, ':title'=>$title, ':content'=>$content));
-                                $update->fetch();
-                                var_dump($update); */
-
-                            }
-                            else {
-                                throw new Exception('Tous les champs ne sont pas remplis !');
-                            }
-                        }  elseif (isset($_POST['delete'])){
-                            deletePost($_GET['edit']);
-                            throw new Exception('Tous les champs ne sont pas remplis !');
-                        }
+                        throw new Exception('Espace réservé à l\'administrateur');
                     }
-
                 break;
         }
     }
